update request body type

+++ base/src/main/java/com/phoenix/base/model/DefaultUserDetails.java
@@ -58,7 +58,7 @@ public class DefaultUserDetails implements UserDetails {
         return userPrincipal.getStatus().equals(ApplicationConstant.USER_DETAILS_STATUS_ENABLED);
     }
 
-    public Long getId() {
+    public Integer getId() {
         return userPrincipal.getId();
     }
